% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_credentials.R
\name{check_user_session}
\alias{check_user_session}
\title{Check if user is still logged in}
\usage{
check_user_session()
}
\value{
State of current user login
}
\description{
Figures out if the user session
is still active and if not it will call
\code{nitrc_login()} to re-establish the session
}
