% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_nitrc_dir.R
\name{download_nitrc_dir}
\alias{download_nitrc_dir}
\title{Download whole image drectory}
\usage{
download_nitrc_dir(session_ID = NULL, subject_ID = NULL, project = NULL,
  scan_type = NULL, zipped = FALSE, verbose = FALSE, error = FALSE)
}
\arguments{
\item{session_ID}{Downoad scan images for this particular session ID}

\item{subject_ID}{Download scan images for this particular subject ID}

\item{project}{The project dataset to whom \code{subject_ID} belongs to.
This is required if subject_ID is specified}

\item{scan_type}{Download scan images just for this particular type of scan}

\item{zipped}{Zip the downloaded files}

\item{verbose}{Should progress be added to download?}

\item{error}{Should function error if download failed?}
}
\value{
List of downloaded file(s) with full paths
}
\description{
Download entire session or just the files from a particular
scan type. If \code{scan_type} is NULL the function will download a zipped
file containing the entire resource folder for that particular \code{subject_ID}
or \code{session_ID}. If \code{scan_type} is specified than depending on the
\code{zipped} the downloaded files will be zipped or not.
}
