% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_this_public.R
\name{is_this_public}
\alias{is_this_public}
\title{Find out if the provided param is part of public data}
\usage{
is_this_public(session_ID = NULL, subject_ID = NULL, project = NULL)
}
\arguments{
\item{session_ID}{the session ID identifier, unique for each individual subject}

\item{subject_ID}{the subject ID identifier}

\item{project}{the NITRC project for which we try to identify if it's public}
}
\value{
logical value signifying if it's part of a public project
}
\description{
Figure out if the provided \code{subject_ID} or
\code{session_ID} belongs to a public project.
}
