% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nitrc_demographics.R
\name{nitrc_demographics}
\alias{nitrc_demographics}
\title{Get/list project demographics}
\usage{
nitrc_demographics(project = NULL)
}
\description{
Retrieves list of all subjects
 demographic data
 @param project is the project fro which
 we request demographics data, if \code{project}
 is NULL, we return all available subjects

 @return Dataframe of demographics data
 @export
}
