% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_credentials.R
\name{nitrc_login}
\alias{nitrc_login}
\title{Login to NITRC}
\usage{
nitrc_login()
}
\value{
boolean indicating if the login was successful
}
\description{
Returns TRUE if NITRC credentials are valid.
}
