% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_credentials.R
\name{query_nitrc}
\alias{query_nitrc}
\title{Perform RCurl operations}
\usage{
query_nitrc(url, jsessionID = NULL)
}
\arguments{
\item{url}{is the URL for the RCurl request parseHTTPHeader}

\item{jsessionID}{value for the JSESSIONID cookie}
}
\description{
Queries NITRC website using RCurl functions
}
