% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nitrc_project.R
\name{read_nitrc_project}
\alias{read_nitrc_project}
\title{Read NITRC project data}
\usage{
read_nitrc_project(project = NULL, jsessionID = NULL)
}
\arguments{
\item{project}{is the project for which
we request the aggregated data, if project
is NULL, we return all available data for all projects}

\item{jsessionID}{value for the JSESSIONID cookie}
}
\value{
Dataframe of project data
}
\description{
Reads in all data associated with a project (demographics and scan data)
}
\examples{
## Read all data for the ixi project
\dontrun{ixi_data <- read_nitrc_project('ixi')}

## Read data for all available projects
\dontrun{nitrc_data <- read_nitrc_project()}
}
