% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_credentials.R
\name{set_credentials}
\alias{set_credentials}
\title{Set NITRC credentials.}
\usage{
set_credentials(username = NULL, password = NULL, error = TRUE)
}
\arguments{
\item{username}{is the NITRC login user name. If \code{NULL} this will
should be provided through the \code{NITRC_WEB_USER} system variable.}

\item{password}{is the NITRC login user password If \code{NULL} this will
should be provided through the \code{NITRC_WEB_PASS} system variable.}

\item{error}{Should this function error if variables are not specifed?}
}
\value{
List of \code{username} and \code{password}
}
\description{
Sets and returns the NITRC credentials. Should error
if it cannot find the environment variables or they are not provided.
}
