% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scan_params.R
\name{get_scan_params}
\alias{get_scan_params}
\title{Get scan parameters data}
\usage{
get_scan_params(session_ID = NULL, scan_type = NULL, project = NULL)
}
\arguments{
\item{session_ID}{the session ID identifier, unique for each individual subject}

\item{scan_type}{the type of scan for which we need to list the acquisition parameters}

\item{project}{optional, signals the project to which the session_ID belongs}
}
\value{
Data frame containing all the data for this particular scan session
}
\description{
Retrieves all the parameters associated with a session ID
}
\examples{
\dontrun{get_scan_params('NITRC_IR_E10469')}
}
