% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_image_sets.R
\name{list_image_sets}
\alias{list_image_sets}
\title{Get NITRC image sets listing}
\usage{
list_image_sets(project = NULL)
}
\value{
Dataframe of NITRC projects
}
\description{
Authenticate and retrieve a list
of all available datasets
}
\examples{
\dontrun{list_image_sets()}
}
