% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_image_sets.R
\name{list_image_sets}
\alias{list_image_sets}
\title{Get NITRC image sets listing}
\usage{
list_image_sets(project = NULL, error = FALSE)
}
\arguments{
\item{project}{optional, called from other functions
in order to figure out if a project is public}

\item{error}{Should function error if httr::GET failed}
}
\value{
Dataframe of NITRC projects
}
\description{
Authenticate and retrieve a list
of all available datasets
}
\examples{
\dontrun{list_image_sets()}
}
