% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nitrc_demographics.R
\name{nitrc_demographics}
\alias{nitrc_demographics}
\title{Get/list project demographics}
\usage{
nitrc_demographics(project = NULL)
}
\arguments{
\item{project}{is the project for which
we request demographics data, if project
is NULL, we return all available subjects}
}
\value{
Dataframe of demographics data
}
\description{
Retrieves list of all subjects demographic data
}
